transformer_ime/requirements.txt
transformers>=4.30.0
datasets>=2.14.0
evaluate>=0.4.0
sacrebleu>=2.3.1
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
mecab-python3>=1.0.6
tensorboard>=2.13.0 